+++ inbm/Changelog.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
  - HSD 15010509095 - INBC fail to return correct Exit Code for difference scenario (-4/-6/-11)
  - HSD 15010982715 - inbm-vision failed to receive query request from cloud
 
+### Added
+- RTC 508495 Support INBC Config Append command
+
 ## 3.0.10 - 2022-03-28
 NOTE: This release has not yet been security tested.